+++ packages/core/RNSentryCocoaTester/RNSentryCocoaTesterTests/RNSentryCocoaTesterTests-Bridging-Header.h
@@ -7,3 +7,6 @@
 #import "RNSentryReplayBreadcrumbConverter.h"
 #import "RNSentryReplayMask.h"
 #import "RNSentryReplayUnmask.h"
+#import "RNSentrySDK.h"
+#import "RNSentryStart.h"
+#import "RNSentryVersion.h"